refactor : use consistent types
change from long to Long, from int to Integer

+++ src/main/java/flab/payment_system/domain/product/controller/ProductController.java
@@ -21,7 +21,7 @@ public class ProductController {
 	private final ProductService productService;
 
 	@GetMapping("/{productId}")
-	public ResponseEntity<ProductDto> getProductDetail(@PathVariable int productId) {
+	public ResponseEntity<ProductDto> getProductDetail(@PathVariable Long productId) {
 		ProductDto productDto = productService.getProductDetail(productId);
 
 		return ResponseEntity.ok().body(productDto);
